requirements.txt
streamlit>=1.31.0
python-dotenv>=1.0.0
google-generativeai>=0.6.1
together>=0.2.11
ollama>=0.2.0